🔧 Foreign Key Naming in Migration Fix

+++ database/migrations/2019_04_29_021400_create_transactions_table.php
@@ -18,8 +18,8 @@ class CreateTransactionsTable extends Migration
             $table->unsignedInteger('customer_id');
             $table->unsignedInteger('partner_id');
             $table->unsignedInteger('plan_id');
-            $table->unsignedBigInteger('agent_id')->nullable();
-            $table->unsignedBigInteger('voucher_id')->nullable();
+            $table->unsignedInteger('agent_id')->nullable();
+            $table->unsignedInteger('voucher_id')->nullable();
             $table->string('insured_relation');
             $table->string('insured_name');
             $table->string('insured_dob');
